+++ src/command/commands/tts.ts
@@ -9,15 +9,23 @@ const command: Command = {
   name: 'tts',
   platforms: [Platform.Twitch],
   run: async (ctx, { services, commandState }) => {
-    return queue.enqueue(async function () {
+    return queue.enqueue(async function (): Promise<boolean> {
       if (!CONFIG.get().openai?.enabled) {
         console.log(`OpenAI is not enabled in !tts command.`);
-        return;
+        return false;
       }
-      if (!ctx.body) return;
+      if (!ctx.body) return false;
       let text = ctx.body;
       text = fixPronunciations(text);
-      await services.openaiService.tts(text, commandState.blunderbotVoice);
+      const res = await services.openaiService.tts(
+        text,
+        commandState.blunderbotVoice
+      );
+      if (!res) {
+        ctx.reply(ctx, `I'm sorry, I couldn't generate audio for that text.`);
+        return false;
+      }
+      return true;
     });
   }
 };